Minor refactoring of scroll_to_bottom implmentation
This just:

- moves the module to the modules folder
- places its call in the DOMContentLoaded event handler
