+++ scripts/src/details.js
@@ -1,10 +1,9 @@
 import manage_details_element from './modules/manage_details_element';
-import scroll_to_bottom from "./scroll_to_bottom";
+import scroll_to_bottom from "./modules/scroll_to_bottom";
 
 const summary_element = document.querySelector('#js-hierarchy-global summary');
 const hierarchy_list = document.querySelector('.hierarchy-global__list');
 
-scroll_to_bottom(hierarchy_list);
 
 summary_element.addEventListener('click', e => {
 
@@ -16,6 +15,7 @@ summary_element.addEventListener('click', e => {
 
 document.addEventListener("DOMContentLoaded", () => {
     manage_details_element();
+    scroll_to_bottom(hierarchy_list);
 });
 
 window.addEventListener("resize", () => {
